added sqlite3_errmsg to all places where sqlite error throws an exception
