+++ dev/constraints.h
@@ -320,7 +320,7 @@ namespace sqlite_orm {
                     case decltype(argument)::nocase: return "NOCASE";
                     case decltype(argument)::rtrim: return "RTRIM";
                 }
-                throw std::domain_error("Invalid collate_argument enum");
+                throw std::system_error(std::make_error_code(orm_error_code::invalid_collate_argument_enum));
             }
         };
