+++ dev/error_code.h
@@ -18,6 +18,8 @@ namespace sqlite_orm {
         cannot_start_a_transaction_within_a_transaction,
         no_active_transaction,
         incorrect_journal_mode_string,
+        incorrect_collate_string,
+        invalid_collate_argument_enum,
     };
     
 }
@@ -51,6 +53,8 @@ namespace sqlite_orm {
                     return "Cannot start a transaction within a transaction";
                 case orm_error_code::no_active_transaction:
                     return "No active transaction";
+                case orm_error_code::invalid_collate_argument_enum:
+                    return "Invalid collate_argument enum";
                 default:
                     return "unknown error";
             }
